+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+**Internal**:
+
+- Switch global rate limiter to a service. ([#4581](https://github.com/getsentry/relay/pull/4581))
+
 ## 25.3.0
 
 **Features**: